The 'install' script will now checkout the haskell mode repository. It will also remove any existing emacs and emacs.d files. In a way not that stupid because it will overwrite the .emacs file anyway, but yes the procedure should be a bit more elaborate. Patches welcome ;-)

+++ emacs/cp_to_home.sh
@@ -1,4 +1,11 @@
 #!/bin/sh
 
+echo "This script is stupid and will remove your .emacs file and .emacs.d folder"
+read -p "ARE YOU SURE? [y/N] " is_sure
+if (! ( [ "$is_sure" = "y" ] || [ "$is_sure" = "Y" ] ) ); then exit 1; fi
+
+rm -rf ~/.emacs ~/.emacs.d
+
 cp emacs ~/.emacs
 cp -r emacs.d ~/.emacs.d
+cd ~/.emacs.d/site-lisp && git clone git://github.com/haskell/haskell-mode.git
